% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean Algorithm.R
\name{euclidean}
\alias{euclidean}
\title{find greatest common divisor of two numbers.}
\usage{
euclidean(first_num, second_num)
}
\arguments{
\item{first_num}{A number.}

\item{second_num}{A number.}
}
\value{
The greatest common divisor of \code{first_num} and \code{second_num}.
}
\description{
find greatest common divisor of two numbers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
