\name{lab3algopackage-package}
\alias{lab3algopackage-package}
\alias{lab3algopackage}
\docType{package}
\title{
\packageTitle{lab3algopackage}
}
\description{
\packageDescription{lab3algopackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3algopackage}
\packageIndices{lab3algopackage}
~~ this package is calculating Euclidean (GCD) of two numbers and Dijkstra (Shotest Path First) from~~ ~~given node~~
}
\author{
\packageAuthor{lab3algopackage}

Maintainer: \packageMaintainer{lab3algopackage}
}
\references{
~~ https://en.wikipedia.org/wiki/Dijkstra's_algorithm ~~
~~ https://en.wikipedia.org/wiki/Euclidean_algorithm ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ dijkstra(wiki_graph, 1)
~~ dijkstra(wiki_graph, 3 ~~
}
