% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra Algorithm.R
\name{dijkstra}
\alias{dijkstra}
\title{find the shortest path between two nodes.}
\usage{
dijkstra(df, init_node)
}
\arguments{
\item{df}{A data frame with two vertices and their weights.}

\item{init_node}{A number as starting node.}
}
\value{
The shortest path vector from \code{init_node} and all other nodes.
}
\description{
find the shortest path between two nodes.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
